requirements.txt
argon2==0.1.10
Flask==3.1.2
flask_sqlalchemy==3.1.1
PyJWT==2.10.1
SQLAlchemy==2.0.39
PyMySQL==1.1.2